:white_check_mark: Rewrite tests
