+++ .gitignore
@@ -1,5 +1,6 @@
 *~
 *.swp
 coverage.html
+package-lock.json
 
 node_modules/
