+++ lib/leanengine.js
@@ -40,7 +40,7 @@ AV.Cloud.HttpsRedirect = function(options) {
 };
 
 AV.Cloud.LeanCloudHeaders = function(options) {
-  return frameworks(require('../middleware/leancloud-heaedrs')(AV)(options), options && options.framework);
+  return frameworks(require('../middleware/leancloud-headers')(AV)(options), options && options.framework);
 };
 
 function createRootRouter(options) {
