+++ test/express/cookie-session-test.js
@@ -1,55 +1,48 @@
-var koa = require('koa');
-var bodyParser = require('koa-bodyparser');
+'use strict';
+
+var express = require('express');
+var bodyParser = require('body-parser');
 var request = require('supertest');
 var should = require('should');
 
-const AV = require('..');
-const appInfo = require('./fixtures/app-info');
+const AV = require('../..');
+const appInfo = require('../fixtures/app-info');
 
-var app = koa();
+var app = express();
 
-app.use(AV.koa());
-app.use(bodyParser());
-app.use(AV.Cloud.CookieSession({framework: 'koa', secret: 'my secret', maxAge: 3600000, fetchUser: true}));
+app.use(AV.express());
+app.use(bodyParser.json());
+app.use(AV.Cloud.CookieSession({secret: 'my secret', maxAge: 3600000, fetchUser: true}));
 
-app.use(function *(next) {
-  try {
-    yield next;
-  } catch (err) {
-    this.status = err.status || 500;
-    this.body = err.message;
-    this.app.emit('error', err, this);
-  }
+app.get('/', function (req, res) {
+  res.send('<p>Hello world</p>');
 });
 
-app.use(function *(next) {
-  var method = this.request.method;
-  var url = this.request.url;
-
-  if (method === 'GET' && url === '/') {
-    this.status = 200;
-    this.body = '<p>Hello world</p>';
-  } else if (method === 'POST' && url === '/login') {
-    return AV.User.logIn(this.request.body.username, this.request.body.password).then( user => {
-      this.response.saveCurrentUser(user);
-      this.response.redirect('/profile');
-    });
-  } else if (method === 'GET' && url === '/profile') {
-    this.status = 200;
-    this.body = this.request.currentUser;
-  } else if (method === 'POST' && url === '/logout') {
-    this.status = 200;
-    this.response.saveCurrentUser(null);
-  } else {
-    yield next;
-  }
+app.post('/login', function(req, res) {
+  AV.User.logIn(req.body.username, req.body.password).then(
+    function(user) {
+      res.saveCurrentUser(user);
+      res.redirect('/profile');
+    },
+    function(error) {
+      res.status = 500;
+      res.send(error);
+    }
+  );
 });
 
-var server = app.listen();
+app.post('/logout', function(req, res) {
+  res.clearCurrentUser();
+  res.send();
+});
 
-describe('webhosting-koa', function() {
+app.get('/profile', function(req, res) {
+  res.send(req.currentUser);
+});
+
+describe('cookie-session', function() {
   it('index', function(done) {
-    request(server).get('/')
+    request(app).get('/')
     .expect(200, function(err, res) {
       res.headers['content-type'].should.be.startWith('text/html');
       res.text.should.be.equal('<p>Hello world</p>');
@@ -58,21 +51,24 @@ describe('webhosting-koa', function() {
   });
 
   it('loign', function(done) {
-    request(server).post('/login')
+    this.timeout(20000);
+
+    request(app).post('/login')
     .send({
       username: 'admin',
       password: 'admin'
     })
     .expect(302, function(err, res) {
+      if (err) return done(err);
       res.headers.location.should.equal('/profile');
       res.headers['set-cookie'][0].indexOf('avos:sess=eyJfdWlkIjoiNTRmZDZhMDNlNGIwNmM0MWUwMGIxZjQwIiwiX3Nlc3Npb25Ub2tlbiI6IncyanJ0a2JlaHAzOG90cW1oYnF1N3liczkifQ==; path=/; expires=').should.equal(0);
       res.headers['set-cookie'][1].indexOf('avos:sess.sig=jMYF3Iwhmw903-K1K12MVdAFOh0; path=/; expires=').should.equal(0);
-      done(err);
+      done();
     });
   });
 
   it('profile', function(done) {
-    request(server).get('/profile')
+    request(app).get('/profile')
     .set('Cookie', 'avos:sess=eyJfdWlkIjoiNTRmZDZhMDNlNGIwNmM0MWUwMGIxZjQwIiwiX3Nlc3Npb25Ub2tlbiI6IncyanJ0a2JlaHAzOG90cW1oYnF1N3liczkifQ==; avos:sess.sig=jMYF3Iwhmw903-K1K12MVdAFOh0')
     .expect(200, function(err, res) {
       should.exist(res.body.objectId);
@@ -82,15 +78,15 @@ describe('webhosting-koa', function() {
   });
 
   it('profile without cookie', function(done) {
-    request(server).get('/profile')
-    .expect(204, function(err, res) {
+    request(app).get('/profile')
+    .expect(200, function(err, res) {
       res.body.should.be.empty();
       done(err);
     });
   });
 
   it('logout', function(done) {
-    request(server).post('/logout')
+    request(app).post('/logout')
     .set('Cookie', 'avos:sess=eyJfdWlkIjoiNTRmZDZhMDNlNGIwNmM0MWUwMGIxZjQwIiwiX3Nlc3Npb25Ub2tlbiI6IncyanJ0a2JlaHAzOG90cW1oYnF1N3liczkifQ==; avos:sess.sig=jMYF3Iwhmw903-K1K12MVdAFOh0')
     .expect(200, function(err, res) {
       res.headers['set-cookie'][0].indexOf('avos:sess=; path=/; expires=').should.equal(0);
