+++ test/express/https-redirect-test.js
@@ -1,10 +1,10 @@
 'use strict';
-var AV = require('..');
+var AV = require('../..');
 var express = require('express');
 var request = require('supertest');
 require('should');
 
-const appInfo = require('./fixtures/app-info');
+const appInfo = require('../fixtures/app-info');
 
 var app = express();
