+++ test/fixtures/app.js
@@ -1,10 +1,10 @@
-const koa = require('koa');
+const Koa = require('koa');
 
 const AV = require('../..');
 
 module.exports = function() {
   if (process.env.FRAMEWORK == 'koa') {
-   var app = koa();
+   var app = new Koa();
    app.use(AV.koa());
    return app.listen();
   } else {
